공통모듈로부터 auth 관심사로 기능 분리
1. Action함수 분리
2. api 분리
3. 공통모듈 api.js, api.test.js 삭제(기능이 없어져서)
4. Reducer 분리
5. 임포트 수정
