--- src/apps/services/__mocks__/api.js
@@ -1,5 +1,0 @@
-export const fetchRestaurants = jest.fn();
-
-export const fetchRestaurantById = jest.fn();
-
-export const authorize = jest.fn();
