+++ src/apps/store/reducer.test.js
@@ -3,8 +3,6 @@ import reducer from './reducer';
 import {
   setLoading,
   setError,
-  setAccessToken,
-  setLoginFields,
 } from './actions';
 
 describe('reducer', () => {
@@ -134,38 +132,6 @@ describe('reducer', () => {
       expect(state.restaurantDetail.isLoading).toBe(false);
       expect(state.restaurantDetail.isError).toBe(true);
       expect(state.restaurantDetail.errorMessage).toBe('에러발생');
-    });
-  });
-
-  describe('setAccessToken', () => {
-    it('set accessToken and initialize status', () => {
-      const initialState = {
-        auth: {
-          accessToken: null,
-          isLoading: false,
-          isError: false,
-          errorMessage: '',
-        },
-      };
-
-      const state = reducer(initialState, setAccessToken('accessToken'));
-
-      expect(state.auth.accessToken).toBe('accessToken');
-    });
-  });
-
-  describe('setLoginFields', () => {
-    it('change login field', () => {
-      const initialState = {
-        loginFields: {
-          email: '',
-          password: '',
-        },
-      };
-
-      const state = reducer(initialState, setLoginFields('email', 'tester@example.com'));
-
-      expect(state.loginFields.email).toBe('tester@example.com');
     });
   });
 });
