+++ src/features/auth/__mocks__/authApi.js
@@ -1,0 +1,3 @@
+export const authorize = jest.fn();
+
+export const login = jest.fn();
